// src/app/models/Vengador.ts
// Generated by https://quicktype.io

export interface Vengador {
    id:              number;
    name:            string;
    realname:        string;
    team:            string;
    firstappearance: string;
    createdby:       string;
    publisher:       string;
    imageurl:        string;
    bio:             string;
}